Update create command
