--- test/commands/hello/index.test.ts
@@ -1,10 +1,0 @@
-import {expect, test} from '@oclif/test'
-
-describe('hello', () => {
-  test
-  .stdout()
-  .command(['hello', 'friend', '--from=oclif'])
-  .it('runs hello cmd', ctx => {
-    expect(ctx.stdout).to.contain('hello friend from oclif!')
-  })
-})
